// frontend/src/core/types/CartItem.ts
export interface CartItem{
    id: number;
    quantity: number;
    user_id: number;
    id_store: number;
    product_id: number;
    product_name: string;
    name_store: string;
    logo_product: string;
    discount: number;
    price_sale: number;
    price: number;
    stock: number;
    attribute1: string;
    attribute2: string;
}